+++ Modules/Biometric/Database/Migrations/2025_06_01_113406_add_card_biometric_employees_table.php
@@ -13,6 +13,10 @@ return new class extends Migration
     public function up(): void
     {
         if (!Schema::hasColumn('biometric_employees', 'card_number')) {
+            if (! Schema::hasTable('biometric_employees')) {
+                return;
+            }
+            
             Schema::table('biometric_employees', function (Blueprint $table) {
                 $table->string('card_number')->nullable()->after('user_id');
             });
@@ -25,6 +29,10 @@ return new class extends Migration
     public function down(): void
     {
         if (Schema::hasColumn('biometric_employees', 'card_number')) {
+            if (! Schema::hasTable('biometric_employees')) {
+                return;
+            }
+            
             Schema::table('biometric_employees', function (Blueprint $table) {
                 $table->dropColumn('card_number');
             });
