+++ Modules/ChattingModule/Database/Migrations/2022_08_21_033729_add_col_to_channel_user_table.php
@@ -17,9 +17,15 @@ class AddColToChannelUserTable extends Migration
             return;
         }
         Schema::table('channel_users', function (Blueprint $table) {
-            $table->boolean('is_read')->default(0);
+            if (! Schema::hasColumn('channel_users', 'is_read')) {
+                $table->boolean('is_read')->default(0)->nullable();
+            }
         });
 
+        if (! Schema::hasTable('channel_lists')) {
+            return;
+        }
+        
         Schema::table('channel_lists', function (Blueprint $table) {
             $table->dropColumn('is_read');
         });
@@ -32,6 +38,10 @@ class AddColToChannelUserTable extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('channel_users')) {
+            return;
+        }
+        
         Schema::table('channel_users', function (Blueprint $table) {
 
         });
