+++ Modules/Payroll/Database/Migrations/2021_09_12_061155_payroll_custom_field_table.php
@@ -18,6 +18,10 @@ return new class extends Migration
                 $table->text('extra_fields')->nullable();
             });
 
+            if (! Schema::hasTable('employee_monthly_salaries')) {
+                return;
+            }
+            
             Schema::table('employee_monthly_salaries', function (Blueprint $table) {
                 $table->enum('allow_generate_payroll', ['yes', 'no'])->default('yes');
             });
@@ -31,10 +35,18 @@ return new class extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('payroll_settings')) {
+            return;
+        }
+        
         Schema::table('payroll_settings', function (Blueprint $table) {
             $table->dropColumn('extra_fields');
         });
 
+        if (! Schema::hasTable('employee_monthly_salaries')) {
+            return;
+        }
+        
         Schema::table('employee_monthly_salaries', function (Blueprint $table) {
             $table->dropColumn('allow_generate_payroll');
         });
