+++ Modules/Recruit/Database/Migrations/2024_02_16_072510_add_company_id_to_application_sources_table.php
@@ -69,6 +69,10 @@ return new class extends Migration {
      */
     public function down(): void
     {
+        if (! Schema::hasTable('application_sources')) {
+            return;
+        }
+        
         Schema::table('application_sources', function (Blueprint $table) {
             $table->dropColumn('company_id');
         });
