+++ Modules/Sms/Database/Migrations/2021_12_13_165351_add_telegram_columns_settings.php
@@ -17,12 +17,22 @@ return new class extends Migration
             return;
         }
         Schema::table('sms_settings', function (Blueprint $table) {
-            $table->boolean('telegram_status')->default(0);
-            $table->string('telegram_bot_token')->nullable();
+            if (! Schema::hasColumn('sms_settings', 'telegram_status')) {
+                $table->boolean('telegram_status')->default(0)->nullable();
+            }
+            if (! Schema::hasColumn('sms_settings', 'telegram_bot_token')) {
+                $table->string('telegram_bot_token')->nullable();
+            }
         });
 
+        if (! Schema::hasTable('users')) {
+            return;
+        }
+        
         Schema::table('users', function (Blueprint $table) {
-            $table->bigInteger('telegram_user_id')->nullable();
+            if (! Schema::hasColumn('users', 'telegram_user_id')) {
+                $table->bigInteger('telegram_user_id')->nullable();
+            }
         });
     }
 
@@ -33,11 +43,19 @@ return new class extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('sms_settings')) {
+            return;
+        }
+        
         Schema::table('sms_settings', function (Blueprint $table) {
             $table->dropColumn(['telegram_status']);
             $table->dropColumn(['telegram_bot_token']);
         });
 
+        if (! Schema::hasTable('users')) {
+            return;
+        }
+        
         Schema::table('users', function (Blueprint $table) {
             $table->dropColumn(['telegram_user_id']);
         });
