+++ Modules/Zoom/Database/Migrations/2020_11_06_042430_add_purchase_code_column_zoom_settings.php
@@ -28,6 +28,10 @@ return new class extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('zoom_setting')) {
+            return;
+        }
+        
         Schema::table('zoom_setting', function (Blueprint $table) {
             $table->dropColumn(['purchase_code', 'supported_until']);
         });
